deployment/requirements.txt
streamlit
pandas
seaborn
matplotlib
pillow
scikit-learn ==1.5.2
xgboost